move conflictPair to type.go, left the start.go only start command related things.add type's unittest type_test.go
do not pass cmd to PreRun. do not destroy the origin function args. check the configure directly whether or not it is flag or configure.
