+++ cmd/exchaind/main.go
@@ -63,7 +63,6 @@ func main() {
 
 	ctx := server.NewDefaultContext()
 
-
 	rootCmd := &cobra.Command{
 		Use:               "exchaind",
 		Short:             "ExChain App Daemon (server)",
@@ -92,7 +91,6 @@ func main() {
 		iaviewerCmd(cdc),
 		subscribeCmd(cdc),
 	)
-
 
 	subFunc := func(logger log.Logger) log.Subscriber {
 		return logevents.NewProvider(logger)
