+++ libs/cosmos-sdk/server/start.go
@@ -70,7 +70,7 @@ func StartCmd(ctx *Context,
 	appStop AppStop,
 	registerRoutesFn func(restServer *lcd.RestServer),
 	registerAppFlagFn func(cmd *cobra.Command),
-	appPreRun func(ctx *Context, cmd *cobra.Command) error,
+	appPreRun func(ctx *Context) error,
 	subFunc func(logger log.Logger) log.Subscriber,
 ) *cobra.Command {
 	cmd := &cobra.Command{
@@ -100,7 +100,7 @@ which accepts a path for the resulting pprof file.
 `,
 		PreRunE: func(cmd *cobra.Command, args []string) error {
 			// app pre run
-			if err := appPreRun(ctx, cmd); err != nil {
+			if err := appPreRun(ctx); err != nil {
 				return err
 			}
 			// set external package flags
